fix(auth): patch roles/permissions bugs

+++ app/Http/Controllers/Dashboard/RoleController.php
@@ -12,6 +12,7 @@ class RoleController extends Controller {
      * Display a listing of the resource.
      */
     public function index() {
+        $this->authorize('roles.view');
         return inertia('admin/roles/Index', [
             'roles' => Role::with(['permissions'])->get(),
         ]);
@@ -21,6 +22,7 @@ class RoleController extends Controller {
      * Show the form for creating a new resource.
      */
     public function create() {
+        $this->authorize('roles.create');
         return inertia(
             'admin/roles/Create',
             ['permissions' => Permission::all()]
@@ -31,12 +33,16 @@ class RoleController extends Controller {
      * Store a newly created resource in storage.
      */
     public function store(Request $request) {
+        // dd($request->all());
+        $this->authorize('roles.create');
         $request->validate([
             'name' => 'required|string|max:255',
+            'name_ar' => 'required|string|max:255',
             'permissions' => 'required|array'
         ]);
         $role = Role::create([
             'name' => $request->name,
+            'name_ar' => $request->name_ar,
             'guard_name' => 'web',
         ]);
         $role->permissions()->sync($request->permissions);
@@ -47,6 +53,7 @@ class RoleController extends Controller {
      * Display the specified resource.
      */
     public function show(Role $role) {
+        $this->authorize('roles.view');
         $role->load('permissions');
         return $role;
     }
@@ -55,6 +62,7 @@ class RoleController extends Controller {
      * Show the form for editing the specified resource.
      */
     public function edit(Role $role) {
+        $this->authorize('roles.edit');
         $role->load('permissions');
         return inertia('admin/roles/Update', [
             'role' => $role,
@@ -66,12 +74,15 @@ class RoleController extends Controller {
      * Update the specified resource in storage.
      */
     public function update(Request $request, Role $role) {
+        $this->authorize('roles.edit');
         $request->validate([
             'name' => 'required|string|max:255',
+            'name_ar' => 'required|string|max:255',
             'permissions' => 'required|array'
         ]);
         $role->update([
             'name' => $request->name,
+            'name_ar' => $request->name_ar,
             'guard_name' => 'web',
         ]);
         $role->permissions()->sync($request->permissions);
@@ -82,6 +93,7 @@ class RoleController extends Controller {
      * Remove the specified resource from storage.
      */
     public function destroy(Role $role) {
+        $this->authorize('roles.delete');
         $role->delete();
         return redirect()->route('roles.index');
     }
